requirements.txt
influxdb
speedtest-cli
requests
pyyaml